Create the TransactionService Class
Create the TransactionService interface. The interface includes methods
for getting a list of a users complete transactions and the
transactions for a single month. In addition, an implementation
TransactionServiceJson is added that reads the list of transactions
from a json resource file. Additional implementations could be added to
read from a database instead.