Create the Controller Class
Create the Rest controller that will serve the rewards reports for a
given user. The controller creates a service dependency based on a
class name given in application.properties. Currently the controller
creates a service that reads data from a json resource file, but if
a service reading from a database is added, the name could be updated
to use that service instead.